// src/app/shared/interfaces/company.ts
export interface CompanyInterface {
    id: number;
    name: string;
    logo: string;
    phone: string;
    email: string;
    about: string;
    file: File;
    facebook: string;
    facebook_fans:number;
    twitter: string;
    twitter_fans:number;
    youtube: string;
    youtube_fans:number;
  }